+++ pycaw/api/endpointvolume/depend.py
@@ -1,0 +1,17 @@
+from ctypes import POINTER, Structure, c_float
+from ctypes.wintypes import BOOL, UINT
+
+from comtypes import GUID
+
+
+class AUDIO_VOLUME_NOTIFICATION_DATA(Structure):
+    _fields_ = [
+        ('guidEventContext', GUID),
+        ('bMuted', BOOL),
+        ('fMasterVolume', c_float),
+        ('nChannels', UINT),
+        ('afChannelVolumes', c_float * 8),
+    ]
+
+
+PAUDIO_VOLUME_NOTIFICATION_DATA = POINTER(AUDIO_VOLUME_NOTIFICATION_DATA)
