+++ pycaw/callbacks.py
@@ -1,0 +1,3 @@
+# TODO
+# implement easy to use callbacks
+# (is already written, but 1 pull at the time)
